+++ src/main/java/com/work/dto/PaymentResultDTO.java
@@ -1,7 +1,9 @@
 package com.work.dto;
 
+import java.math.BigDecimal;
+
 public class PaymentResultDTO {
     public int installmentsPaid;
-    public double totalPaid;
+    public BigDecimal totalPaid;
     public boolean loanFullyPaid;
 }
